moves a few files
